update requirement and separate the test bksf into another file
