+++ requirements.txt
@@ -6,3 +6,4 @@ psutil
 jsonschema
 pyobjc-core; sys_platform == 'darwin'
 pyobjc-framework-Cocoa; sys_platform == 'darwin'
+networkx
